This is a very large update for WIN32.  Virtually all headers have been changed to accomodate WIN32 linking peculiarities.
Several .pro files and .pri files have been changed also.  
This change requires a ./makeallclean.sh followed by a .makeall.sh.

The local.pri file also needs to be updated.  There are two new 
entried needed:

MYSQLPATH   = /usr/include/mysql
QWT3D       = ../../qwtplot3d-qt4

+++ gui/us_associations_gui.cpp
@@ -168,7 +168,7 @@ void US_AssociationsGui::set_component( int index, int row, int col )
    w              = L->itemAt( 0 )->widget();
    QwtCounter*  c = dynamic_cast< QwtCounter* >( w );
 
-   c->setValue( fabs( as->stoichs[ index ] ) ); 
+   c->setValue( fabs( (double)as->stoichs[ index ] ) ); 
 }
 
 void US_AssociationsGui::changed( int row, int col )
